Language/03_text_generation/requirements.txt
numpy==1.23.5
pandas==2.0.1
tensorflow_cpu==2.12.0
kaggle==1.5.13